+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "vscode-didact" extension will be documented in this file.
 
+## 0.1.16
+
+ - TBD
+
 ## 0.1.15
 
  - Adding "Validate Didact File" command to provide diagnostic validation